feat: classes generated with grammar #21

+++ exam.domain/src/main/java/org/domain/model/SolutionWeight.java
@@ -7,19 +7,19 @@ import javax.persistence.Embeddable;
 public class SolutionWeight {
 
     @Column(name = "solution_weight", nullable = false)
-    private final int value;
+    private final double value;
 
 
     protected SolutionWeight(){
         this.value = 0;
     }
 
-    public SolutionWeight(int value) {
+    public SolutionWeight(double value) {
         // for ORM
         this.value = value;
     }
 
-    public int value() {
+    public double value() {
         return value;
     }
 }
